fix: ProfileSection 이동 및 컴포넌트 분리 #46
