+++ src/components/ProfileSection/index.ts
@@ -1,0 +1,1 @@
+import { default as ProfileSection } from "./ProfileSection";
